Avoid calling `origin_ids` when we already have real ids.

Performance:
```
# before vs after, small recordset - bigger difference for bigger recordsets
In [1]: %timeit self.ids
189 ns ± 1.5 ns per loop
90.6 ns ± 0.499 ns per loop 

In [2]: %timeit self._origin
395 ns ± 5.43 ns per loop
35.6 ns ± 0.254 ns per loop
```

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr