@robodoo 2025-03-25
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/203242.png)](https://mergebot.odoo.com/odoo/odoo/pull/203242)