@rco-odoo 2025-03-25
Updated performance: for a single record
```
In [1]: %timeit self.ids
353 ns ± 2.27 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
153 ns ± 0.104 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)

In [2]: %timeit self._origin
647 ns ± 3.23 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
77.9 ns ± 0.0982 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)
```
For 37 records
```
In [4]: %timeit partners.ids
1.52 µs ± 31.8 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
354 ns ± 3.26 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

In [5]: %timeit partners._origin
1.8 µs ± 35.1 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
242 ns ± 0.939 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```
In summary, the observed speedup factor is between 2.3 and 4.3 for `records.ids`, and between 7.4 and 8.3 for `records._origin`.