@robodoo 2025-03-25
@kmagusiak @rco-odoo 'ci/security' failed on this reviewed PR.