@rco-odoo 2025-03-25
@robodoo override=ci/security

Use of `getattr()` in refactored code: it's the same code put in a different place.  As a reminder, we use `getattr` on a record id to retrieve its `origin`; `getattr()` is used to avoid some `try`/`except` (because `int` don't have `origin`).